third_party/blink/web_tests/external/wpt/css/css-display/display-math-on-non-mathml-elements-expected.txt

This is a testharness.js-based test.
PASS computed display on <DIV style="display: math;">
PASS computed display on <DIV style="display: inline-math;">
FAIL computed display on <DIV style="display: inline math;"> assert_equals: expected "inline" but got "block"
PASS computed display on <DIV style="display: block math;">
PASS computed display on <svg style="display: math;">
PASS computed display on <svg style="display: inline-math;">
PASS computed display on <svg style="display: inline math;">
FAIL computed display on <svg style="display: block math;"> assert_equals: expected "block" but got "inline"
Harness: the test ran to completion.
